Added a Vote comman
Added a vote command
Made the code more organized
Fixed the missing thing in config
Make config better
Changed the Readme.md
