--- us/scandic/simpleGeyserVL/QueuedVote.java
@@ -1,17 +1,0 @@
-package us.scandic.simpleGeyserVL;
-
-import java.util.UUID;
-
-public class QueuedVote
-{
-	UUID PlayerID;
-	String Service;
-	long Timestamp;
-	
-	public QueuedVote(UUID uid, String service, long ts)
-	{
-		PlayerID = uid;
-		Service = service;
-		Timestamp = ts;
-	}
-}
